Updated version of the website and Api
Fixed the CSS file, created the schema_database for the charity events, and also included Node Modules which contain the package_lock.json, and package.json files. The website is now functioning well, the search engine is effectively retrieving data from the database. The register page will be intergrated in assessment 3.